--- src/abi/PriceFeed.ts
@@ -1,765 +1,0 @@
-export const PriceFeedABI = [
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: '_rbac',
-        type: 'address',
-      },
-      {
-        internalType: 'address',
-        name: '_dataFabric',
-        type: 'address',
-      },
-      {
-        internalType: 'address',
-        name: 'pythContract',
-        type: 'address',
-      },
-    ],
-    stateMutability: 'nonpayable',
-    type: 'constructor',
-  },
-  {
-    inputs: [],
-    name: 'GracePeriodNotOver',
-    type: 'error',
-  },
-  {
-    inputs: [],
-    name: 'InvalidMarketDecimals',
-    type: 'error',
-  },
-  {
-    inputs: [],
-    name: 'InvalidMarketId',
-    type: 'error',
-  },
-  {
-    inputs: [],
-    name: 'InvalidPrice',
-    type: 'error',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'string',
-        name: 'errorMsg',
-        type: 'string',
-      },
-    ],
-    name: 'InvalidRole',
-    type: 'error',
-  },
-  {
-    inputs: [],
-    name: 'InvalidToken',
-    type: 'error',
-  },
-  {
-    inputs: [],
-    name: 'SequencerDown',
-    type: 'error',
-  },
-  {
-    inputs: [],
-    name: 'ZeroAddress',
-    type: 'error',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: true,
-        internalType: 'address',
-        name: 'feedAddress',
-        type: 'address',
-      },
-    ],
-    name: 'ChainlinkSequencerUptimeFeedUpdated',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: true,
-        internalType: 'bytes32',
-        name: 'marketId',
-        type: 'bytes32',
-      },
-      {
-        indexed: false,
-        internalType: 'bytes32',
-        name: 'priceId',
-        type: 'bytes32',
-      },
-    ],
-    name: 'MarketFeedUpdatedPyth',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: false,
-        internalType: 'uint256',
-        name: 'feesPaid',
-        type: 'uint256',
-      },
-    ],
-    name: 'PythPriceUpdated',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: false,
-        internalType: 'bool',
-        name: 'status',
-        type: 'bool',
-      },
-    ],
-    name: 'SequencerUptimeStatusUpdated',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: true,
-        internalType: 'address',
-        name: 'tokenAddress',
-        type: 'address',
-      },
-      {
-        indexed: false,
-        internalType: 'bytes32',
-        name: 'priceId',
-        type: 'bytes32',
-      },
-    ],
-    name: 'TokenFeedUpdatedPyth',
-    type: 'event',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'bytes32',
-        name: 'marketId',
-        type: 'bytes32',
-      },
-      {
-        internalType: 'uint256',
-        name: 'marketAmount',
-        type: 'uint256',
-      },
-      {
-        internalType: 'address',
-        name: 'tokenAddress',
-        type: 'address',
-      },
-    ],
-    name: 'convertMarketToToken',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: 'tokenAmount',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'bytes32',
-        name: 'marketId',
-        type: 'bytes32',
-      },
-      {
-        internalType: 'uint256',
-        name: 'marketAmount',
-        type: 'uint256',
-      },
-      {
-        internalType: 'address',
-        name: 'tokenAddress',
-        type: 'address',
-      },
-    ],
-    name: 'convertMarketToTokenSecondary',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: 'tokenAmount',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'bytes32',
-        name: 'marketId',
-        type: 'bytes32',
-      },
-      {
-        internalType: 'uint256',
-        name: 'amount',
-        type: 'uint256',
-      },
-    ],
-    name: 'convertMarketToUsd',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: 'amountUsd',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: 'tokenA',
-        type: 'address',
-      },
-      {
-        internalType: 'uint256',
-        name: 'amountA',
-        type: 'uint256',
-      },
-      {
-        internalType: 'address',
-        name: 'tokenB',
-        type: 'address',
-      },
-    ],
-    name: 'convertTokenToToken',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: 'amountB',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: 'tokenA',
-        type: 'address',
-      },
-      {
-        internalType: 'uint256',
-        name: 'amountA',
-        type: 'uint256',
-      },
-      {
-        internalType: 'address',
-        name: 'tokenB',
-        type: 'address',
-      },
-    ],
-    name: 'convertTokenToTokenSecondary',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: 'amountB',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: 'tokenAddress',
-        type: 'address',
-      },
-      {
-        internalType: 'uint256',
-        name: 'amount',
-        type: 'uint256',
-      },
-    ],
-    name: 'convertTokenToUsd',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: 'amountUsd',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'dataFabric',
-    outputs: [
-      {
-        internalType: 'contract DataFabric',
-        name: '',
-        type: 'address',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'bytes32',
-        name: 'marketId',
-        type: 'bytes32',
-      },
-    ],
-    name: 'getLatestPriceMarket',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: 'marketPriceUsd',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: 'tokenAddress',
-        type: 'address',
-      },
-    ],
-    name: 'getLatestPriceToken',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: 'tokenPriceUsd',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'bytes32',
-        name: 'marketId',
-        type: 'bytes32',
-      },
-    ],
-    name: 'getMarketPricePrimary',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: 'marketPriceUsd',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'bytes32',
-        name: 'marketId',
-        type: 'bytes32',
-      },
-    ],
-    name: 'getMarketPricePyth',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: 'marketPriceUsd',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'bytes32',
-        name: 'marketId',
-        type: 'bytes32',
-      },
-    ],
-    name: 'getMarketPricePythEMA',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: 'marketPriceUsd',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'bytes32',
-        name: 'marketId',
-        type: 'bytes32',
-      },
-    ],
-    name: 'getMarketPriceSecondary',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: 'marketPriceUsd',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: 'tokenAddress',
-        type: 'address',
-      },
-    ],
-    name: 'getTokenPricePrimary',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: 'tokenPriceUsd',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: 'tokenAddress',
-        type: 'address',
-      },
-    ],
-    name: 'getTokenPricePyth',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: 'tokenPriceUsd',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: 'tokenAddress',
-        type: 'address',
-      },
-    ],
-    name: 'getTokenPricePythEMA',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: 'tokenPriceUsd',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: 'tokenAddress',
-        type: 'address',
-      },
-    ],
-    name: 'getTokenPriceSecondary',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: 'tokenPriceUsd',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'getValidTimePeriod',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: '',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'bytes32',
-        name: '',
-        type: 'bytes32',
-      },
-    ],
-    name: 'marketToPythPriceId',
-    outputs: [
-      {
-        internalType: 'bytes32',
-        name: '',
-        type: 'bytes32',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'pyth',
-    outputs: [
-      {
-        internalType: 'contract IPyth',
-        name: '',
-        type: 'address',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'rbac',
-    outputs: [
-      {
-        internalType: 'contract IRBAC',
-        name: '',
-        type: 'address',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: '',
-        type: 'address',
-      },
-    ],
-    name: 'tokenToPythPriceId',
-    outputs: [
-      {
-        internalType: 'bytes32',
-        name: '',
-        type: 'bytes32',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'bytes32',
-        name: 'marketId',
-        type: 'bytes32',
-      },
-      {
-        internalType: 'bytes[]',
-        name: 'priceUpdateData',
-        type: 'bytes[]',
-      },
-    ],
-    name: 'updateAndGetMarketPrice',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: 'marketPriceUsd',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'bytes32',
-        name: 'marketId',
-        type: 'bytes32',
-      },
-      {
-        internalType: 'bytes[]',
-        name: 'priceUpdateData',
-        type: 'bytes[]',
-      },
-    ],
-    name: 'updateAndGetMarketPriceEMA',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: 'marketPriceUsd',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: 'tokenAddress',
-        type: 'address',
-      },
-      {
-        internalType: 'bytes[]',
-        name: 'priceUpdateData',
-        type: 'bytes[]',
-      },
-    ],
-    name: 'updateAndGetTokenPrice',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: 'tokenPriceUsd',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: 'tokenAddress',
-        type: 'address',
-      },
-      {
-        internalType: 'bytes[]',
-        name: 'priceUpdateData',
-        type: 'bytes[]',
-      },
-    ],
-    name: 'updateAndGetTokenPriceEMA',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: 'tokenPriceUsd',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: '_sequencerUptimeFeed',
-        type: 'address',
-      },
-    ],
-    name: 'updateChainlinkSequencerUptimeFeed',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'bytes32',
-        name: 'marketId',
-        type: 'bytes32',
-      },
-      {
-        internalType: 'bytes32',
-        name: 'priceId',
-        type: 'bytes32',
-      },
-    ],
-    name: 'updateMarketFeedPyth',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'bytes[]',
-        name: 'priceUpdateData',
-        type: 'bytes[]',
-      },
-    ],
-    name: 'updatePythPrice',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'bool',
-        name: 'status',
-        type: 'bool',
-      },
-    ],
-    name: 'updateSequencerUptimeStatus',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: 'tokenAddress',
-        type: 'address',
-      },
-      {
-        internalType: 'bytes32',
-        name: 'priceId',
-        type: 'bytes32',
-      },
-    ],
-    name: 'updateTokenFeedPyth',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    stateMutability: 'payable',
-    type: 'receive',
-  },
-] as const;
