--- src/abi/arb-sepolia/ParifiUtils.ts
@@ -1,197 +1,0 @@
-export const ParifiUtilsABI = [
-  {
-    inputs: [
-      { internalType: 'address', name: '_orderManager', type: 'address' },
-    ],
-    stateMutability: 'nonpayable',
-    type: 'constructor',
-  },
-  {
-    inputs: [{ internalType: 'string', name: 'errorMsg', type: 'string' }],
-    name: 'InvalidRole',
-    type: 'error',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: true, internalType: 'bytes32', name: 'id', type: 'bytes32' },
-      {
-        indexed: false,
-        internalType: 'uint256',
-        name: 'position',
-        type: 'uint256',
-      },
-    ],
-    name: 'BatchExecuteFailed',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: true, internalType: 'bytes32', name: 'id', type: 'bytes32' },
-      {
-        indexed: false,
-        internalType: 'uint256',
-        name: 'position',
-        type: 'uint256',
-      },
-    ],
-    name: 'BatchExecuteSuccess',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: false,
-        internalType: 'string',
-        name: 'message',
-        type: 'string',
-      },
-    ],
-    name: 'Log',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: false, internalType: 'bytes', name: 'message', type: 'bytes' },
-    ],
-    name: 'LogBytes',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      {
-        indexed: true,
-        internalType: 'address',
-        name: 'previousOwner',
-        type: 'address',
-      },
-      {
-        indexed: true,
-        internalType: 'address',
-        name: 'newOwner',
-        type: 'address',
-      },
-    ],
-    name: 'OwnershipTransferred',
-    type: 'event',
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: false, internalType: 'bool', name: 'enabled', type: 'bool' },
-    ],
-    name: 'PublicAccess',
-    type: 'event',
-  },
-  {
-    inputs: [],
-    name: 'allowPublicAccess',
-    outputs: [{ internalType: 'bool', name: '', type: 'bool' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        components: [
-          { internalType: 'bytes32', name: 'id', type: 'bytes32' },
-          { internalType: 'bytes[]', name: 'priceUpdateData', type: 'bytes[]' },
-        ],
-        internalType: 'struct ParifiUtils.BatchExecute[]',
-        name: 'batchedPositions',
-        type: 'tuple[]',
-      },
-    ],
-    name: 'batchLiquidatePositions',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        components: [
-          { internalType: 'bytes32', name: 'id', type: 'bytes32' },
-          { internalType: 'bytes[]', name: 'priceUpdateData', type: 'bytes[]' },
-        ],
-        internalType: 'struct ParifiUtils.BatchExecute[]',
-        name: 'batchedOrders',
-        type: 'tuple[]',
-      },
-    ],
-    name: 'batchSettleOrders',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      { internalType: 'bytes32', name: 'id', type: 'bytes32' },
-      { internalType: 'bytes[]', name: 'priceUpdateData', type: 'bytes[]' },
-    ],
-    name: 'liquidatePosition',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'orderManager',
-    outputs: [
-      { internalType: 'contract IOrderManager', name: '', type: 'address' },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'owner',
-    outputs: [{ internalType: 'address', name: '', type: 'address' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'rbac',
-    outputs: [{ internalType: 'contract IRBAC', name: '', type: 'address' }],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'renounceOwnership',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      { internalType: 'bytes32', name: 'id', type: 'bytes32' },
-      { internalType: 'bytes[]', name: 'priceUpdateData', type: 'bytes[]' },
-    ],
-    name: 'settleOrder',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      { internalType: 'bool', name: '_allowPublicAccess', type: 'bool' },
-    ],
-    name: 'togglePublicAccess',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [{ internalType: 'address', name: 'newOwner', type: 'address' }],
-    name: 'transferOwnership',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-] as const;
