--- src/abi/arb-sepolia/index.ts
@@ -1,23 +1,0 @@
-import { OrderManagerABI } from './OrderManager';
-import { PriceFeedABI } from './PriceFeed';
-import { ParifiVaultABI } from './ParifiVault';
-import { ParifiForwarderABI } from './ParifiForwarder';
-import { DataFabricABI } from './DataFabric';
-import { FeeManagerABI } from './FeeManager';
-import { AdapterABI } from './Adapter';
-import { ParifiUtilsABI } from './ParifiUtils';
-import { RBACABI } from './RBAC';
-import { SubgraphHelperABI } from './SubgraphHelper';
-
-export {
-  FeeManagerABI,
-  AdapterABI,
-  OrderManagerABI,
-  DataFabricABI,
-  PriceFeedABI,
-  ParifiVaultABI,
-  ParifiForwarderABI,
-  ParifiUtilsABI,
-  RBACABI,
-  SubgraphHelperABI,
-};
