+++ src/abis/parifi/FairsaleAbi.ts
